Updates mode in open to be enum not string, fixes tests.
Also tried to fix some Mac compilation errors, but I have
yet to try it myself on Mac OS.

Backed out change to tasks-fifo.c
Other "maybe I'm not using it" changes from the code review -
I am using the changes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/ferguson@19415 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/include/qio.h
@@ -105,8 +105,8 @@ extern ssize_t qio_mmap_chunk_iobufs;
  */
 err_t qio_readv(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t* num_read);
 err_t qio_writev(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t* num_written);
-err_t qio_preadv(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t seek_to_offset, int64_t* num_read);
-err_t qio_pwritev(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t seek_to_offset, int64_t* num_written);
+err_t qio_preadv(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t seek_to_offset, ssize_t* num_read);
+err_t qio_pwritev(fd_t fd, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t seek_to_offset, ssize_t* num_written);
 
 err_t qio_freadv(FILE* fp, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t* num_read);
 err_t qio_fwritev(FILE* fp, qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end, ssize_t* num_written);
