+++ runtime/src/sys.c
@@ -1180,7 +1180,7 @@ err_t sys_listen(fd_t sockfd, int backlog)
   return err_out;
 }
 
-err_t sys_recv(fd_t sockfd, void* buf, size_t len, int flags, int64_t* num_recvd_out)
+err_t sys_recv(fd_t sockfd, void* buf, size_t len, int flags, ssize_t* num_recvd_out)
 {
   ssize_t got;
   err_t err_out;
@@ -1199,7 +1199,7 @@ err_t sys_recv(fd_t sockfd, void* buf, size_t len, int flags, int64_t* num_recvd
   return err_out;
 }
 
-err_t sys_recvfrom(fd_t sockfd, void* buf, size_t len, int flags, sys_sockaddr_t* src_addr_out, int64_t* num_recvd_out)
+err_t sys_recvfrom(fd_t sockfd, void* buf, size_t len, int flags, sys_sockaddr_t* src_addr_out, ssize_t* num_recvd_out)
 {
   ssize_t got;
   err_t err_out;
@@ -1218,7 +1218,7 @@ err_t sys_recvfrom(fd_t sockfd, void* buf, size_t len, int flags, sys_sockaddr_t
   return err_out;
 }
 
-err_t sys_recvmsg(fd_t sockfd, struct msghdr *msg, int flags, int64_t* num_recvd_out)
+err_t sys_recvmsg(fd_t sockfd, struct msghdr *msg, int flags, ssize_t* num_recvd_out)
 
 {
   ssize_t got;
@@ -1240,7 +1240,7 @@ err_t sys_recvmsg(fd_t sockfd, struct msghdr *msg, int flags, int64_t* num_recvd
   return err_out;
 }
 
-err_t sys_send(fd_t sockfd, const void* buf, int64_t len, int flags, int64_t* num_sent_out)
+err_t sys_send(fd_t sockfd, const void* buf, int64_t len, int flags, ssize_t* num_sent_out)
 {
   ssize_t sent;
   err_t err_out;
@@ -1259,7 +1259,7 @@ err_t sys_send(fd_t sockfd, const void* buf, int64_t len, int flags, int64_t* nu
   return err_out;
 }
 
-err_t sys_sendto(fd_t sockfd, const void* buf, int64_t len, int flags, const sys_sockaddr_t* dest_addr, int64_t* num_sent_out)
+err_t sys_sendto(fd_t sockfd, const void* buf, int64_t len, int flags, const sys_sockaddr_t* dest_addr, ssize_t* num_sent_out)
 {
   ssize_t sent;
   err_t err_out;
@@ -1278,7 +1278,7 @@ err_t sys_sendto(fd_t sockfd, const void* buf, int64_t len, int flags, const sys
   return err_out;
 }
 
-err_t sys_sendmsg(fd_t sockfd, const struct msghdr *msg, int flags, int64_t* num_sent_out)
+err_t sys_sendmsg(fd_t sockfd, const struct msghdr *msg, int flags, ssize_t* num_sent_out)
 {
   ssize_t sent;
   err_t err_out;
